refactor: migrate LoreEntry cards
