+++ src/components/cards/CardContainer.vue
@@ -5,7 +5,7 @@
 		with-options
 		@edit="showForm"
 		@delete="confirmDelete"
-		@move="showFolderTree"
+		@move="moveLoreEntry"
 		@dragstart="onDragStart"
 	>
 		<!-- Dynamic Card content component -->
@@ -16,22 +16,24 @@
 </template>
 
 <script lang="ts" setup>
+import { useRepo } from "pinia-orm";
 import { computed, defineAsyncComponent, ref } from "vue";
 import { CustomMIMEType, startDrag } from "@/composables/dragAndDrop";
-import { CardTypes, getText } from "@/core/model/cards";
+import { LoreEntry } from "@/core/models";
+import { LoreEntryRepo } from "@/core/repositories";
 import { t as $t } from "@/core/translation";
 import utilities from "@/core/utils/functions";
-import { useCardsStore } from "@/store/cards";
 import { useGlobalConfirmDialog } from "@/store/confirmDialog";
 import { useSidePanel } from "@/store/sidePanel";
 import BaseCard from "./BaseCard.vue";
 import CardDragImage from "./CardDragImage.vue";
 
-const props = defineProps<{ itemData: CardTypes; draggable?: boolean }>();
+const props = defineProps<{ itemData: LoreEntry; draggable?: boolean }>();
 
 const refDragImage = ref<HTMLElement | null>();
 
-const cardsStore = useCardsStore();
+const loreEntryRepo = useRepo(LoreEntryRepo);
+
 const sidePanelStore = useSidePanel();
 const { showConfirmDialog } = useGlobalConfirmDialog();
 
@@ -45,23 +47,23 @@ function onDragStart(e: DragEvent) {
 }
 
 function showForm() {
-	sidePanelStore.newFileEditForm(props.itemData.id, cardsStore.currentFolder);
+	sidePanelStore.showLoreEntryForm("edit", props.itemData);
 }
 
-function showFolderTree() {
-	sidePanelStore.newFolderTree(props.itemData, cardsStore.currentFolder);
+function moveLoreEntry() {
+	sidePanelStore.showFolderTree(props.itemData);
 }
 
 function confirmDelete() {
 	showConfirmDialog({
-		title: $t(`dialogs.delete${utilities.capitalize(props.itemData._category)}`),
-		message: $t(`dialogs.deleteConfirm`) + `"${getText(props.itemData)}" ?`,
-		confirmAction: () => cardsStore.deleteCard(props.itemData),
+		title: $t(`dialogs.delete${utilities.capitalize(props.itemData.category)}`),
+		message: $t(`dialogs.deleteConfirm`) + `"${props.itemData.getText()}" ?`,
+		confirmAction: () => loreEntryRepo.delete(props.itemData.id),
 	});
 }
 
 const contentComponent = computed(() => {
-	const componentName = "Content" + utilities.capitalize(props.itemData._category);
+	const componentName = "Content" + utilities.capitalize(props.itemData.category);
 	return defineAsyncComponent({
 		loader: () => import(`./files/content/${componentName}.vue`),
 	});
