+++ src/components/cards/files/content/ContentLocation.vue
@@ -15,7 +15,7 @@
 <script lang="ts" setup>
 import TagList from "@/components/cards/tags/TagList.vue";
 import MarkdownView from "@/components/common/MarkdownView.vue";
-import { Location } from "@/core/model/cards";
+import { Location } from "@/core/models";
 
 defineProps<{
 	itemData: Location;
