+++ src/components/cards/files/forms/FormLocation.vue
@@ -13,7 +13,7 @@
 <script lang="ts" setup>
 import { computed } from "vue";
 import TagListPanel from "@/components/cards/tags/TagListPanel.vue";
-import { Location } from "@/core/model/cards";
+import { Location } from "@/core/models";
 import { t as $t } from "@/core/translation";
 import { required } from "@/core/validationRules";
